feat: 实现删除标签

+++ src/Action/DestroyTrait.php
@@ -2,7 +2,9 @@
 
 namespace Miaoxing\Admin\Action;
 
+use Miaoxing\Plugin\BaseModelV2;
 use Miaoxing\Plugin\Service\Convention;
+use Miaoxing\Plugin\Service\Request;
 use Wei\RetTrait;
 
 /**
@@ -16,8 +18,18 @@ trait DestroyTrait
     {
         $model = $this->convention->createModel($this)->findOneById($req['id']);
 
+        $ret = $this->beforeDestroy($req, $model);
+        if ($ret['code'] !== 1) {
+            return $ret;
+        }
+
         $model->destroy();
 
+        return $this->suc();
+    }
+
+    protected function beforeDestroy(Request $req, BaseModelV2 $model)
+    {
         return $this->suc();
     }
 }
